basic refactor, splitting tabs and utils
