+++ config.py
@@ -1,0 +1,4 @@
+from load_config import load_config
+
+
+config = load_config()
