+++ get_date.py
@@ -1,0 +1,6 @@
+import datetime
+
+
+def get_date():
+    now = datetime.datetime.now()
+    return now.strftime("%Y-%m-%d_%H-%M-%S")
