+++ load_config.py
@@ -1,0 +1,6 @@
+import json
+
+
+def load_config():
+    with open('config.json', 'r') as infile:
+        return json.load(infile)
