+++ model_manager.py
@@ -1,0 +1,5 @@
+from BarkModelManager import BarkModelManager
+from config import config
+
+
+model_manager = BarkModelManager(config)
