+++ models/clone.sh
@@ -1,3 +1,5 @@
 # Clone one of bark model forks within this directory
-git clone https://github.com/rsxdalv/bark.git
-git clone https://github.com/rsxdalv/tortoise-tts.git
+git clone https://github.com/rsxdalv/bark.git bark
+
+# Clone one of tortoise model forks within this directory
+git clone https://github.com/rsxdalv/tortoise-tts.git tortoise
